Move Coord2d to separate file
