+++ adventofcode/d17.py
@@ -2,8 +2,9 @@ import heapq
 import logging
 from dataclasses import dataclass, field
 
+from adventofcode.tooling.coordinates import Coord2d
 from adventofcode.tooling.directions import CardinalDirection as Dir
-from adventofcode.tooling.map import Coord2d, Map2d
+from adventofcode.tooling.map import Map2d
 
 _logger = logging.getLogger(__name__)
