+++ adventofcode/d18.py
@@ -3,8 +3,8 @@ import logging
 from collections.abc import Iterable
 from dataclasses import dataclass, field
 
+from adventofcode.tooling.coordinates import Coord2d
 from adventofcode.tooling.directions import CardinalDirection as Dir
-from adventofcode.tooling.map import Coord2d
 
 _logger = logging.getLogger(__name__)
